Refactor to individual files
