+++ libs/portals/core/src/lib/messages.ts
@@ -105,6 +105,10 @@ export const m = defineMessages({
     id: 'portals:dashboard-islandis',
     defaultMessage: 'Stjórnborð Ísland.is',
   },
+  moduleSwitcher: {
+    id: 'portals:module-switcher',
+    defaultMessage: 'Stjórnborðs valmynd',
+  },
   openModuleSwitcher: {
     id: 'portals:open-module-switcher',
     defaultMessage: 'Opna stjórnborðs valmynd',
